Wrap existing code into build tag webhdfs
